requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-core==0.1.1
python-dotenv==1.0.0
openai==1.2.3
pypdf==3.16.4
psycopg2-binary==2.9.9
pgvector==0.2.3
tiktoken==0.5.1
google-generativeai==0.3.1